Rename marketplace login to xpkg login
Ditto logout.
