--- cmd/crank/marketplace/marketplace.go
@@ -1,25 +1,0 @@
-/*
-Copyright 2023 The Crossplane Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-// Package marketplace contains commands for interacting with the Upbound
-// Marketplace.
-package marketplace
-
-// Cmd contains commands for interacting with the Upbound Marketplace.
-type Cmd struct {
-	Login  loginCmd  `cmd:"" help:"Login to the Upbound Marketplace."`
-	Logout logoutCmd `cmd:"" help:"Logout of the Upbound Marketplace."`
-}
