+++ cmd/crank/xpkg/xpkg.go
@@ -20,8 +20,10 @@ package xpkg
 // Cmd contains commands for interacting with xpkgs.
 // TODO(lsviben) add the rest of the commands from up (batch, xpextract).
 type Cmd struct {
-	Build buildCmd `cmd:"" help:"Build a package, by default from the current directory."`
-	Push  pushCmd  `cmd:"" help:"Push Crossplane packages."`
+	Build  buildCmd  `cmd:"" help:"Build a package, by default from the current directory."`
+	Push   pushCmd   `cmd:"" help:"Push Crossplane packages."`
+	Login  loginCmd  `cmd:"" help:"Login to the default package registry (xpkg.upbound.io)."`
+	Logout logoutCmd `cmd:"" help:"Logout of the default package registry (xpkg.upbound.io)."`
 }
 
 // Help prints out the help for the xpkg command.
